requirements.txt
numpy==1.23.5
torch~=2.2.1 --index-url https://download.pytorch.org/whl/cu121
keyboard>=0.13.5
gymnasium>=0.28.1
pyqt5>=5.15.0
mlagents-env==1.2.0.dev0 
mlagents==1.2.0.dev0 
protobuf>=3.20.3